Rebalance refactor (#37)
* WIP rebalance protocol to move data around

* rebalance WIP, need to error handle and retry

* (WIP) rebalance and distribute, begin refactor

* fix neuron event handler

* (WIP) rebalance and distribute, start monitor protocol

* (WIP) update rebalance protocol, ping and retry as a modular block

* sane import pruning, refactor to reuse primitives for rebalance

* update distribute to reuse primitives

* Fix typo in validator/store/store_broadband

* Move reward vector fulfillment to delegated function for reuse

* Add rewards vector to store_broadband function

* Call apply_reward_scores in store_broadband

* Fix missing imports and argument typo

* (WIP) api fixes

* api fixes to use store/retrieve primitives, import fixes, black

* api working with new extracted rewards

* update README with bonding curve

* update readme again

* fix verify by choosing correct fn per protocol

* update event handler wrppaer

* don't support py 3.11

* bulk fix for refactor and rebalance

* add event updates to broadband, rm uid from rebal, update req to bt642

* bulk update... need more testing

* revert handler

* fix event handler in run

* fix rebal event passed arg

* remove round-robin validator

* update reamde

---------
