+++ storage/validator/reward.py
@@ -18,6 +18,7 @@
 
 
 import torch
+import typing
 import numpy as np
 import bittensor as bt
 
@@ -232,12 +233,12 @@ def apply_reward_scores(
     )
     bt.logging.debug(f"Scaled rewards: {scaled_rewards}")
 
-    # Compute forward pass rewards, assumes followup_uids and answer_uids are mutually exclusive.
+    # Compute forward pass rewards
     # shape: [ metagraph.n ]
     scattered_rewards: torch.FloatTensor = self.moving_averaged_scores.scatter(
         0, torch.tensor(uids).to(self.device), scaled_rewards
     ).to(self.device)
-    bt.logging.debug(f"Scattered rewards: {scattered_rewards}")
+    bt.logging.trace(f"Scattered rewards: {scattered_rewards}")
 
     # Update moving_averaged_scores with rewards produced by this step.
     # shape: [ metagraph.n ]
@@ -245,4 +246,69 @@ def apply_reward_scores(
     self.moving_averaged_scores: torch.FloatTensor = alpha * scattered_rewards + (
         1 - alpha
     ) * self.moving_averaged_scores.to(self.device)
-    bt.logging.debug(f"Updated moving avg scores: {self.moving_averaged_scores}")
+    bt.logging.trace(f"Updated moving avg scores: {self.moving_averaged_scores}")
+
+
+from bittensor import Synapse
+
+from .verify import (
+    verify_store_with_seed,
+    verify_challenge_with_seed,
+    verify_retrieve_with_seed,
+)
+from .database import add_metadata_to_hotkey
+from .bonding import update_statistics, get_tier_factor
+from .event import EventSchema
+from storage.protocol import Store, Retrieve
+
+import sys
+from pprint import pformat
+
+
+async def create_reward_vector(
+    self,
+    synapse: typing.Union[Store, Retrieve],
+    rewards: torch.FloatTensor,
+    uids: list[int],
+    responses: list[Synapse],
+    event: EventSchema,
+    callback: callable,
+    fail_callback: callable,
+):
+    if isinstance(synapse, Store):
+        verify_fn = verify_store_with_seed
+    elif isinstance(synapse, Retrieve):
+        verify_fn = verify_retrieve_with_seed
+    elif isinstance(synapse, Challenge):
+        verify_fn = verify_challenge_with_seed
+    else:
+        raise ValueError(f"Invalid synapse type: {type(synapse)}")
+
+    for idx, (uid, response) in enumerate(zip(uids, responses)):
+        # Verify the commitment
+        hotkey = self.metagraph.hotkeys[uid]
+        success = verify_fn(response)
+        if success:
+            bt.logging.debug(f"Successfully verified store commitment from UID: {uid}")
+            await callback(hotkey, idx, uid, response)
+        else:
+            bt.logging.error(f"Failed to verify store commitment from UID: {uid}")
+            fail_callback(uid)
+
+        # Update the storage statistics
+        await update_statistics(
+            ss58_address=hotkey,
+            success=success,
+            task_type="store",
+            database=self.database,
+        )
+
+        # Apply reward for this store
+        tier_factor = await get_tier_factor(hotkey, self.database)
+        rewards[idx] = 1.0 * tier_factor if success else 0.0
+
+        event.successful.append(success)
+        event.uids.append(uid)
+        event.completion_times.append(response.dendrite.process_time)
+        event.task_status_messages.append(response.dendrite.status_message)
+        event.task_status_codes.append(response.dendrite.status_code)
